requirements.txt
# first apt install python3-tk
numpy
torchvision
matplotlib
scipy
torch
vispy
opencv_python
opencv_contrib_python
Pillow
PyYAML
datetime
yaml
argparse
